+++ routes/add.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+
+router.get("/", (req, res) => {
+  res.render("addPackage.ejs", { message: req.flash("message") });
+});
+module.exports = router;
